Move vendor neutral OSS components
