+++ bindings/nacos/nacos_test.go
