+++ bindings/rocketmq/settings.go
