+++ bindings/rocketmq/settings_test.go
@@ -19,7 +19,7 @@ import (
 	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/require"
 
-	"github.com/dapr/components-contrib/bindings/alicloud/rocketmq"
+	"github.com/dapr/components-contrib/bindings/rocketmq"
 )
 
 func TestSettingsDecode(t *testing.T) { //nolint:paralleltest
