+++ tests/certification/bindings/dubbo/components/dubbo.yaml
