+++ tests/certification/bindings/nacos/components/nacos.yaml
