+++ tests/certification/bindings/nacos/config.yaml
